refactor(status): colocate readiness checks with their domain files
Each sub-resource readiness function (knative, logging, route, eventing,
volumes) now lives next to the status builder it relates to instead of
in the central capp.go. Inlines the single-use buildCappConditions
wrapper and replaces hardcoded condition strings with library constants.
